make some functions happy
