+++ src/commandlog.c
@@ -169,6 +169,147 @@ void commandlogPushCurrentCommand(client *c, struct serverCommand *cmd) {
     commandlogPushEntryIfNeeded(c, argv, argc, net_output_bytes_curr_cmd, COMMANDLOG_TYPE_LARGE_REPLY);
 }
 
+// bigKeyInfo [string | set | hash | list | zset | xstream | all]
+void bigkeyInfoCommand(client *c) {
+    bool isDisplayAll = 0;
+    bool isDisplayString = 0;
+    bool isDisplaySet = 0;
+    bool isDisplayHash = 0;
+    bool isDisplayList = 0;
+    bool isDisplayZset = 0;
+    long totalLine = 0;
+
+    long count_string = listLength(server.string_bigkey_info);
+    long count_list = listLength(server.list_bigkey_info);
+    long count_hash = listLength(server.hash_bigkey_info);
+    long count_set = listLength(server.set_bigkey_info);
+    long count_zset = listLength(server.zset_bigkey_info);
+
+    if (c->argc == 1) {
+        isDisplayAll = 1;
+    } else {
+        int j = 1;
+        while (j < c->argc) {
+            char *opt = c->argv[j]->ptr;
+            if (!strcasecmp(opt, "string")) {
+                isDisplayString = 1;
+            } else if (!strcasecmp(opt, "set")) {
+                isDisplaySet = 1;
+            } else if (!strcasecmp(opt, "hash")) {
+                isDisplayHash = 1;
+            } else if (!strcasecmp(opt, "list")) {
+                isDisplayList = 1;
+            } else if (!strcasecmp(opt, "zset")) {
+                isDisplayZset = 1;
+            } else if (!strcasecmp(opt, "all")) {
+                isDisplayAll = 1;
+                break;
+            } else {
+                addReplyErrorFormat(c, "Unsupported option %s", opt);
+                return;
+            }
+            j++;
+        }
+    }
+
+    if (server.big_key_output < count_string) {
+        count_string = server.big_key_output;
+    }
+    if (server.big_key_output < count_list) {
+        count_list = server.big_key_output;
+    }
+    if (server.big_key_output < count_hash) {
+        count_hash = server.big_key_output;
+    }
+    if (server.big_key_output < count_set) {
+        count_set = server.big_key_output;
+    }
+    if (server.big_key_output < count_zset) {
+        count_zset = server.big_key_output;
+    }
+
+    if (isDisplayAll | isDisplayString) {
+        totalLine += count_string;
+    }
+    if (isDisplayAll | isDisplaySet) {
+        totalLine += count_set;
+    }
+    if (isDisplayAll | isDisplayHash) {
+        totalLine += count_hash;
+    }
+    if (isDisplayAll | isDisplayList) {
+        totalLine += count_list;
+    }
+    if (isDisplayAll | isDisplayZset) {
+        totalLine += count_zset;
+    }
+
+    addReplyArrayLen(c, totalLine);
+
+    listIter li;
+    listNode *ln;
+
+    if (isDisplayAll | isDisplayString) {
+        listRewindTail(server.string_bigkey_info, &li);
+        while (count_string--) {
+            ln = listNext(&li);
+            bigkeyEntry *bke = ln->value;
+            addReplyArrayLen(c, 3);
+            addReplyLongLong(c, bke->value);
+            addReplyBulkCBuffer(c, bke->key->ptr, sdslen(bke->key->ptr));
+            addReplyBulkCString(c, "string");
+        }
+    }
+
+    if (isDisplayAll | isDisplayList) {
+        listRewindTail(server.list_bigkey_info, &li);
+        while (count_list--) {
+            ln = listNext(&li);
+            bigkeyEntry *bke = ln->value;
+            addReplyArrayLen(c, 3);
+            addReplyLongLong(c, bke->value);
+            addReplyBulkCBuffer(c, bke->key->ptr, sdslen(bke->key->ptr));
+            addReplyBulkCString(c, "list");
+        }
+    }
+
+    if (isDisplayAll | isDisplayHash) {
+        listRewindTail(server.hash_bigkey_info, &li);
+        while (count_hash--) {
+            ln = listNext(&li);
+            bigkeyEntry *bke = ln->value;
+            addReplyArrayLen(c, 3);
+            addReplyLongLong(c, bke->value);
+            addReplyBulkCBuffer(c, bke->key->ptr, sdslen(bke->key->ptr));
+            addReplyBulkCString(c, "hash");
+        }
+    }
+
+    if (isDisplayAll | isDisplaySet) {
+        listRewindTail(server.set_bigkey_info, &li);
+        while (count_set--) {
+            ln = listNext(&li);
+            bigkeyEntry *bke = ln->value;
+            addReplyArrayLen(c, 3);
+            addReplyLongLong(c, bke->value);
+            addReplyBulkCBuffer(c, bke->key->ptr, sdslen(bke->key->ptr));
+            addReplyBulkCString(c, "set");
+        }
+    }
+
+    if (isDisplayAll | isDisplayZset) {
+        listRewindTail(server.list_bigkey_info, &li);
+        while (count_zset--) {
+            ln = listNext(&li);
+            bigkeyEntry *bke = ln->value;
+            addReplyArrayLen(c, 3);
+            addReplyLongLong(c, bke->value);
+            addReplyBulkCBuffer(c, bke->key->ptr, sdslen(bke->key->ptr));
+            addReplyBulkCString(c, "Sorted Set");
+        }
+    }
+}
+
 /* The SLOWLOG command. Implements all the subcommands needed to handle the
  * slow log. */
 void slowlogCommand(client *c) {
