+++ src/server.c
@@ -38,7 +38,6 @@
 #include "cluster_slot_stats.h"
 #include "cluster_migrateslots.h"
 #include "commandlog.h"
-#include "keyinfo.h"
 #include "bio.h"
 #include "latency.h"
 #include "mt19937-64.h"
@@ -2990,6 +2989,109 @@ serverDb *createDatabaseIfNeeded(int id) {
         server.db[id] = createDatabase(id);
     }
     return server.db[id];
+}
+
+void removeNodeFromList(dataType previous_type, robj *keyobj) {
+    list *process_list = NULL;
+    if (previous_type == STRING_TYPE) {
+        process_list = server.string_bigkey_info;
+    } else if (previous_type == LIST_TYPE) {
+        process_list = server.list_bigkey_info;
+    } else if (previous_type == HASH_TYPE) {
+        process_list = server.hash_bigkey_info;
+    } else if (previous_type == SET_TYPE) {
+        process_list = server.set_bigkey_info;
+    } else if (previous_type == SORTED_SET_TYPE) {
+        process_list = server.zset_bigkey_info;
+    } else {
+        // TO DO later
+    }
+
+    listIter li;
+    listNode *ln;
+    listRewind(process_list, &li);
+    while ((ln = listNext(&li)) != NULL) {
+        bigkeyEntry *node = ln->value;
+        if (sdscmp(node->key->ptr, keyobj->ptr) != 0) continue;
+        decrRefCount(node->key);
+        listDelNode(process_list, ln);
+        zfree(node);
+    }
+}
+
+void addNodeToList(dataType new_type, robj *keyobj, long curr) {
+    list *process_list = NULL;
+    if (new_type == STRING_TYPE) {
+        process_list = server.string_bigkey_info;
+    } else if (new_type == LIST_TYPE) {
+        process_list = server.list_bigkey_info;
+    } else if (new_type == HASH_TYPE) {
+        process_list = server.hash_bigkey_info;
+    } else if (new_type == SET_TYPE) {
+        process_list = server.set_bigkey_info;
+    } else if (new_type == SORTED_SET_TYPE) {
+        process_list = server.zset_bigkey_info;
+    } else {
+        // TO DO later
+    }
+
+    bigkeyEntry *bke = zmalloc(sizeof(bigkeyEntry));
+    incrRefCount(keyobj);
+
+    bke->value = curr;
+    bke->key = keyobj;
+
+    if (listLength(process_list) == 0) {
+        listAddNodeHead(process_list, bke);
+    } else {
+        listIter li;
+        listNode *ln;
+        listRewind(process_list, &li);
+        int isInsert = 0;
+        while ((ln = listNext(&li)) != NULL) {
+            bigkeyEntry *node = ln->value;
+            if (curr > node->value) continue;
+            listInsertNode(process_list, ln, bke, 0);
+            isInsert = 1;
+            break;
+        }
+        if (!isInsert) listAddNodeTail(process_list, bke);
+    }
+}
+
+
+void updateBigKeyList(robj *keyobj, long previous, long curr, dataType type) {
+    if (type == STRING_TYPE) {
+        if (previous < server.string_memory_use && curr < server.string_memory_use) {
+            return;
+        } else if (previous >= server.string_memory_use && curr < server.string_memory_use) {
+            removeNodeFromList(type, keyobj);
+        } else if (previous < server.string_memory_use && curr >= server.string_memory_use) {
+            addNodeToList(type, keyobj, curr);
+        } else if (previous >= server.string_memory_use && curr >= server.string_memory_use) {
+            removeNodeFromList(type, keyobj);
+            addNodeToList(type, keyobj, curr);
+        }
+    } else {
+        if (previous < server.big_key_number_element && curr < server.big_key_number_element) {
+            return;
+        } else if (previous >= server.big_key_number_element && curr < server.big_key_number_element) {
+            removeNodeFromList(type, keyobj);
+        } else if (previous < server.big_key_number_element && curr >= server.big_key_number_element) {
+            addNodeToList(type, keyobj, curr);
+        } else if (previous >= server.big_key_number_element && curr >= server.big_key_number_element) {
+            removeNodeFromList(type, keyobj);
+            addNodeToList(type, keyobj, curr);
+        }
+    }
+}
+
+void bigkeyListInit(void) {
+    server.string_bigkey_info = listCreate();
+    server.list_bigkey_info = listCreate();
+    server.hash_bigkey_info = listCreate();
+    server.set_bigkey_info = listCreate();
+    server.zset_bigkey_info = listCreate();
 }
 
 void resetDBKeySizeArray(serverDb *db) {
